+++ mastermind.rb
@@ -1,87 +1,14 @@
-class Peg
-  COLORS = [ "yellow", "green", "purple", "orange", "blue", "brown"]
-  attr_accessor :color
-  def initialize (color)
-    @color = color
-  end
-end
-
-class Row 
-
-  attr_accessor :slots
-  def initialize
-    @slots = []
-  end
-
-  def populate_row ( colors )
-    colors.each do | color |
-      @slots << Peg.new( color )
-    end
-  end
-
-  def to_s
-    colors = @slots.map { |slot| slot.color }
-    colors.join(", ")
-  end
-
-end
-
-class Board
-  attr_accessor :solution_row, :guesses
-  def initialize
-    @solution_row = Row.new
-    @guesses = []
-  end
-
-  def create_solution ( solution )
-    @solution_row.populate_row ( solution )
-  end
-
-  def create_autogenerate_solution 
-    solution = []
-    4.times do
-      solution << Peg::COLORS.sample
-    end
-    create_solution( solution )
-  end
-
-  def add_guess_row( guess )
-    guess_row = Row.new
-    guess_row.populate_row(guess)
-    @guesses << guess_row
-  end
-end
-
-class Player
-  def make_a_guess
-    puts "What is your guess?"
-    guess = gets.chomp.split(",")
-    guess.each{ |color| color.strip! }
-    guess
-  end
-end
-
-class Human < Player
-  def make_a_guess
-    puts "What is your guess?"
-    guess = gets.chomp.split(",")
-    guess.each{ |color| color.strip! }
-    guess
-  end
-end
-
-class Computer < Player
-  def make_a_guess
-  end
-end
-
+require_relative './computer.rb'
+require_relative './human.rb'
+require_relative './peg.rb'
+require_relative './row.rb'
+require_relative './board.rb'
 
 class Mastermind
-
   def initialize
     @board = Board.new
-    @master = Player.new
-    @guesser = Player.new
+    @codemaker = Computer.new
+    @codebreaker = Human.new
   end
 
   def instructions
@@ -89,12 +16,27 @@ class Mastermind
     puts "if incorrect, make another"
   end
 
+  def render
+    puts @board.to_s
+  end
+
   def play
     instructions
 
+    @board.create_autogenerate_solution
 
+    loop do
+      @board.add_guess_row(@codebreaker.make_a_guess)
 
-    @board.add_guess_row(@guesser.make_a_guess)
+      if check_win?
+        puts "You won!"
+        break
+      else
+        @board.add_response(respond_to_guess)
+      end
+
+      render
+    end
   end
 
   def check_win?
@@ -102,11 +44,11 @@ class Mastermind
   end
 
   def respond_to_guess
-    guess = @board.guesses.last
-    solution = @board.solution_row
+    guess = @board.guesses.last.slots
+    solution = @board.solution_row.slots
     response = []
-    remainder_guess = {}
-    remainder_solution = {}
+    guess_colors = {}
+    solution_colors = {}
 
     (0..3).each do |index|
       if guess[index] == solution[index]
@@ -115,23 +57,23 @@ class Mastermind
         guess_color = guess[index].color
         solution_color = solution[index].color
 
-        if remainder_guess.keys.include?(guess_color)
-          remainder_guess[guess_color] += 1
+        if guess_colors.keys.include?(guess_color)
+          guess_colors[guess_color] += 1
         else
-          remainder_guess[guess_color] = 1
+          guess_colors[guess_color] = 1
         end
 
-        if remainder_solution.keys.include?(solution_color)
-          remainder_solution[solution_color] += 1
+        if solution_colors.keys.include?(solution_color)
+          solution_colors[solution_color] += 1
         else
-          remainder_solution[solution_color] = 1
+          solution_colors[solution_color] = 1
         end
       end
     end
 
-    remainder_guess.each do |color, number|
-      if remainder_solution.keys.include?(color)
-        color_matches = [remainder_solution[color], remainder_guess[color]].min
+    guess_colors.each do |color, number|
+      if solution_colors.keys.include?(color)
+        color_matches = [solution_colors[color], guess_colors[color]].min
         color_matches.times do
           response << "color match"
         end
@@ -143,11 +85,7 @@ class Mastermind
 end
 
 test_game = Mastermind.new
-test_board = Board.new
-test_board.create_autogenerate_solution()
-puts test_board.solution_row.to_s
-test_board.add_guess_row(Row.new.populate_row(["yellow", "blue", "orange", "green"]))
-puts test_board.guesses.last.to_s
+test_game.play
 
 # the_mind = Mastermind.new
 # the_mind.play
