+++ peg.rb
@@ -1,0 +1,7 @@
+class Peg
+  COLORS = [ "yellow", "green", "purple", "orange", "blue", "brown"]
+  attr_accessor :color
+  def initialize (color)
+    @color = color
+  end
+end
