+++ player.rb
@@ -1,0 +1,4 @@
+class Player
+  def make_a_guess
+  end
+end
